LPS-189113 Remove service from PLOLanguageOverrideProvider, create PLOLanguageOverrideProviderHelper and apply to usages
