+++ modules/apps/portal-language/portal-language-override-service/src/main/java/com/liferay/portal/language/override/internal/PLOEntryModelListener.java
@@ -101,13 +101,13 @@ public class PLOEntryModelListener
 		MethodType methodType, PLOEntry ploEntry) {
 
 		if (methodType == MethodType.ADD) {
-			_ploLanguageOverrideProvider.add(ploEntry);
+			_ploLanguageOverrideProviderHelper.add(ploEntry);
 		}
 		else if (methodType == MethodType.REMOVE) {
-			_ploLanguageOverrideProvider.remove(ploEntry);
+			_ploLanguageOverrideProviderHelper.remove(ploEntry);
 		}
 		else if (methodType == MethodType.UPDATE) {
-			_ploLanguageOverrideProvider.update(ploEntry);
+			_ploLanguageOverrideProviderHelper.update(ploEntry);
 		}
 	}
 
@@ -122,7 +122,8 @@ public class PLOEntryModelListener
 	private ClusterExecutor _clusterExecutor;
 
 	@Reference
-	private PLOLanguageOverrideProvider _ploLanguageOverrideProvider;
+	private PLOLanguageOverrideProviderHelper
+		_ploLanguageOverrideProviderHelper;
 
 	private enum MethodType {
